requirements.txt
asgiref==3.2.9
Django==2.2
django-cors-headers==3.4.0
djangorestframework==3.7.7
gunicorn==20.0.4
mysqlclient==1.4.6
